apps/ai-worker/requirements.txt
# PyTorch with CUDA 12.1
--index-url https://download.pytorch.org/whl/cu121
torch==2.4.1
torchvision==0.19.1

# Diffusion models
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0

# HuggingFace Hub (model downloads)
huggingface-hub>=0.23.0

# Image processing
Pillow>=10.0.0

# Voice generation (Coqui TTS)
coqui-tts>=0.24.0

# Music & SFX generation (AudioCraft)
audiocraft>=1.3.0

# Video enhancement
opencv-python-headless>=4.8.0
realesrgan>=0.3.0
basicsr>=1.4.0

# Frame interpolation (optional — falls back to OpenCV optical flow)
# rife-ncnn-vulkan-python
